requirements.txt
# LangChain Core
langchain>=0.1.0
langchain-anthropic>=0.1.0
langchain-openai>=0.1.0
langchain-pinecone>=0.1.0
langchain-community>=0.0.20
langchain-cohere>=0.1.0

# Vector Stores
pinecone-client>=3.2.2,<4.0.0

# Reranking
cohere>=5.1.4,<6.0.0

# PPT Processing
python-pptx==0.6.23
Pillow==10.2.0

# OCR (optional)
pytesseract==0.3.10
opencv-python==4.9.0.80

# BM25
rank-bm25==0.2.2

# Web Framework
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6

# Streamlit Frontend
streamlit==1.31.0
streamlit-chat==0.1.1

# Utilities
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
tiktoken==0.5.2

# Rate Limiting
ratelimit==2.2.1
tenacity==8.2.3

# Monitoring & Logging
prometheus-client==0.19.0
loguru==0.7.2

# Caching
redis==5.0.1
diskcache==5.6.3

# Testing
pytest==8.0.0
pytest-asyncio==0.23.3
pytest-mock==3.12.0

# Development
black==24.1.1
ruff==0.1.14
pre-commit==3.6.0
